Changes in EventController and EventService, added correct objects as method arguments usage, minor changes

+++ src/main/java/org/semester/service/EventServiceImpl.java
@@ -2,8 +2,9 @@ package org.semester.service;
 
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
-import org.semester.dto.EventDto;
-import org.semester.dto.UserDto;
+import org.semester.dto.eventDto.EventDto;
+import org.semester.dto.eventDto.OnCreateEventDto;
+import org.semester.dto.userDto.UserDto;
 import org.semester.entity.Event;
 import org.semester.entity.EventImage;
 import org.semester.entity.User;
@@ -12,7 +13,6 @@ import org.semester.mappers.UserMapper;
 import org.semester.repository.EventImageRepository;
 import org.semester.repository.EventRepository;
 import org.semester.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
@@ -42,9 +42,12 @@ public class EventServiceImpl implements EventService {
     private EventMapper eventMapper;
 
     @Override
-    public Event addEvent(Event event) {
-        return eventRepository.saveAndFlush(event);
+    public EventDto addEvent(OnCreateEventDto onCreateEventDto, String email) {
+        User foundUser = userRepository.findByEmail(email);
+        Event newEvent = eventRepository.saveAndFlush(eventMapper.getEventEntity(onCreateEventDto, foundUser));
+        return eventMapper.getEventDto(newEvent);
     }
+
     @Override
     public List<EventDto> getEvents(Integer page) {
         return eventRepository.findAll(PageRequest.of(page, PAGE_SIZE)).map(
@@ -80,10 +83,15 @@ public class EventServiceImpl implements EventService {
 
     @Override
     @Transactional
-    public void deleteEvent(Long id) {
-        Event event = eventRepository.findById(id).orElseThrow();
+    public Boolean deleteEvent(Long id) {
+        Optional<Event> eventOptional = eventRepository.findById(id);
+        if (eventOptional.isEmpty()) {
+            return false;
+        }
+        Event event = eventOptional.get();
         deleteImageFromDisk(event.getEventImages());
         eventRepository.delete(event);
+        return true;
     }
 
     @Override
@@ -112,11 +120,20 @@ public class EventServiceImpl implements EventService {
 
     @Override
     @Transactional
-    public void deleteImage(Long id) {
-        Event event = eventRepository.findById(id).orElseThrow();
+    public Boolean deleteImage(Long id) {
+        Optional<Event> eventOptional = eventRepository.findById(id);
+        if (eventOptional.isEmpty()) {
+            return false;
+        }
+        Event event = eventOptional.get();
         List<EventImage> images = event.getEventImages();
-        deleteImageFromDisk(images);
+        try {
+            deleteImageFromDisk(images);
+        } catch (RuntimeException e) {
+            return false;
+        }
         eventImageRepository.deleteByEventId(id);
+        return true;
     }
 
     @Override
